+++ src/services/notifications/index.ts
@@ -1,4 +1,4 @@
-import type { OAuthClient } from '../../types';
+import type { AuthHttpClient } from '../../types';
 import httpClient from '../http-client';
 import { decamelizeKeys } from '../../http/utils';
 import notifications from './notifications';
@@ -8,7 +8,7 @@ import { NOTIFICATIONS_BASE } from '../../constants';
 import { NotificationSettingsServices, NotificationsService } from './types';
 
 export const notificationsService = (
-  httpWithAuth: OAuthClient
+  httpWithAuth: AuthHttpClient
 ): NotificationsService &
   ReturnType<typeof health> & {
     settings: NotificationSettingsServices;